api/requirements.txt
argon2-cffi==21.3.0
beautifulsoup4==4.10.0
daphne==3.0.2
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-ipware==4.0.2
django-redis==5.2.0
django-user-agents==0.4.0
djangorestframework==3.13.1
freezegun==1.2.1
psycopg2-binary==2.9.3
python-decouple==3.6
requests==2.27.1
